Euphs.Bot

Description

Provides a framework to build bots connecting to heim. This module is still a bit unstable, and between API changes, it will break a lot.

Structures and type synonyms

data Bot

The main Bot data structure.

Constructors

Bot

Fields

botConnection :: Connection

Websocket connection to heim.

packetCount :: PacketID

The packet counter

botAgent :: BotAgent

The Bot agent given from the server

botRoom :: String

The room the bot currently is in.

botName :: String

Initial bot nick

startTime :: UTCTime

Time at which the bot was started

botFun :: BotFunctions

Custom bot functions

sideThreads :: TVar [ThreadId]

An experimental way to keep track of the threads spawned

logHandle :: Handle

Logging handle

evtQueue :: TQueue EuphEvent

Queue of reply events

roomPW :: Maybe String

Room password

closeVar :: TChan ()

data BotFunctions

Custom Bot functions

Constructors

BotFunctions

Fields

eventsHook :: EuphEvent -> Net ()

Main event loop. Every event not handled by the bot, calls this function.

dcHook :: Maybe (IO ())

Special actions to run in a disconnect, for cleanup.

helpShortHook :: Maybe ([String] -> Net String)

A short !help description

helpLongHook :: Maybe ([String] -> Net String)

A long !help botName description

type Net = ReaderT Bot IO

The monad transformer stack, for handling all the bot events

type BotAgent = TMVar UserData

The mutable bot agent

type PacketID = TVar Int

A mutable counter for packet ids

Main Functions

bot :: BotFunctions -> IO ()

The main bot call function. When this action ends, the bot is closed.

botWithOpts :: BotFunctions -> Opts -> IO ()

Function for calling the bot with custom options, using a getOpts call.

closeBot :: Net ()

Function for closing the bot from inside the Net monad

disconnect :: BotFunctions -> IO ()

Function for closing off the bot.

emptyBot :: BotFunctions

Empty bot

getBotAgent :: Net UserData

An easier way to read the current bot agent

sendPacket :: EuphCommand -> Net EuphEvent

Sends an Heim command, and fetches the reply.

getBotConfig :: FromJSON a => Opts -> IO (Maybe a)

Interface to the YAML parser, using the Bot commanline interface

tellLog :: Text -> Net ()

Logging information to the log handle.

sendReply :: MessageData -> String -> Net EuphEvent

Convenience function for sendPacket $ Send str (msgID m)