Euphs.Easy

Description

An Easier way of making a bot, that has basic needs, like an exact word match for the command, | and providing an easy way to also have an auto generated help.

Documentation

easyBot :: [EasyBotFun] -> IO ()

The main function call.

type EasyBotFun = (Command, (EasyFun, ShortHelp, LongHelp))

A structure for defining a command.

type Command = String

A '!command' string, to define the trigger of the action.

type ShortHelp = String

Short help about the command implemented

type LongHelp = String

Long help about the command implemented

type EasyFun = String -> Net String

A function with the rest of the command as input, and the output as a direct reply.