Euphs.Events

Description

Module Declaring all the possible events the bot can receive during a connection, while putting them into a structure.

Documentation

data EuphEvent

The main Event structure, every JSON reply gets parsed into this structure, and then passed to the custom bot function.

Constructors

PingEvent

Fields

pingTime :: Integer

The sent time

nextTime :: Integer

Next time a ping will be sent

WhoReply

Fields

idReply :: Int

ID based on the command sent

users :: [UserData]

List of sessions connected

LogReply

Fields

idReply :: Int

ID based on the command sent

messages :: [MessageData]

List of messages requested

SnapshotEvent

Fields

identity :: String

Id of the agent logged

sessionID :: String

Global ID of this session

version :: String

Server's version identifier

users :: [UserData]

List of sessions connected

messages :: [MessageData]

List of messages requested

SendEvent

Fields

msgData :: MessageData

Messsage sent from an user

SendReply

Fields

idReply :: Int

ID based on the command sent

msgData :: MessageData

Messsage sent from an user

NickEvent

Fields

userData :: UserData

New user agent

fromNick :: String

Previous nick

NickReply

Fields

idReply :: Int

ID based on the command sent

userData :: UserData

New user agent

fromNick :: String

Previous nick

JoinEvent

Fields

userData :: UserData

New user agent

PartEvent

Fields

userData :: UserData

New user agent

HelloEvent

Fields

userData :: UserData

New user agent

privateRoom :: Bool

Indicating whether a room is private or not.

version :: String

Server's version identifier

AuthReply

Fields

idReply :: Int

ID based on the command sent

success :: Bool

Whether the authentication was successful

reason :: String

Reason of denied auth

BounceEvent

Fields

reason :: String

Reason of denied auth

decodePacket :: ByteString -> Either String EuphEvent

Function to decode a network packet to an Euphorian event

encodePacket :: SentCommand -> ByteString

Function to encode an Euphorian command to a network packet

matchIdReply :: Int -> EuphEvent -> Bool

Function to match a Reply with the command