Euphs.Options

Description

This module defines the type Opts and gives a function to parse commandline arguments.

Documentation

data Opts

The main options structure.

Constructors

Opts

Fields

heimHost :: String

Heim instance hoster

heimPort :: Int

The port to connect to

useSSL :: Bool

Whether to use wss or ws

roomList :: String

A list of initial rooms separated by whitespace in the format of name[-pw]

showHelp :: Bool

Print help on startup

logTarget :: FilePath

The log file to write to. Fall back to stdout when empty

botAccount :: FilePath

Email and password for logging into the account

botNick :: String

The nick the Bot will be using.

config :: FilePath

Config file for specific bot options

parseOpts :: Opts -> OptsList -> [String] -> IO (Opts, [String])

Function to parse commandline Options

type OptsList = [OptDescr (Opts -> Opts)]

Type synonym for the list of options to the bot

defaults :: Opts

Default options

options :: OptsList

List of options available

getOpts :: Opts -> OptsList -> IO Opts

Parses options from the commandline, using the Opts given in as a default to be overridden