Euphs.Types

Description

This module defines the Types and structure of the packets incoming. | Also look at the official api.

Documentation

type MessageID = String

A type synonym for Snowflake.

data UserData

Structure representing an agent, user or bot. Also called SessionView, in the api.

Constructors

UserData

Fields

userID :: String

ID of the user

name :: String

Current nick of the user

serverID :: String

The id of the server

serverEra :: String

The era of the server

sessionID :: String

ID of the session, unique across all sessions, globally

isStaff :: Bool

Whether the session, is a member of the staff

isManager :: Bool

Whether the session, is a manager of the room

data MessageData

Structure representing a Message incoming from the server.

Constructors

MessageData

Fields

timeRecieved :: Integer

msgID :: MessageID

parentMsg :: String

prevEditID :: String

sender :: UserData

contentMsg :: String

encryptKey :: String

edited :: Maybe Integer

deleted :: Maybe Integer

truncated :: Bool